[crashtracker] Remove the old dataformat
