+++ crashtracker-ffi/src/crash_info/api.rs
@@ -1,7 +1,7 @@
 // Copyright 2024-Present Datadog, Inc. https://www.datadoghq.com/
 // SPDX-License-Identifier: Apache-2.0
 
-use datadog_crashtracker::rfc5_crash_info::CrashInfo;
+use datadog_crashtracker::CrashInfo;
 use ddcommon::Endpoint;
 use ddcommon_ffi::{wrap_with_void_ffi_result, Handle, ToInner, VoidResult};
 use function_name::named;
