+++ crashtracker-ffi/src/crash_info/stackframe.rs
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use ::function_name::named;
-use datadog_crashtracker::rfc5_crash_info::{BuildIdType, FileType, StackFrame};
+use datadog_crashtracker::{BuildIdType, FileType, StackFrame};
 use ddcommon_ffi::{
     slice::AsBytes, wrap_with_void_ffi_result, CharSlice, Handle, Result, ToInner, VoidResult,
 };
