+++ crashtracker/src/crash_info/builder.rs
