+++ crashtracker/src/crash_info/error_data.rs
@@ -3,7 +3,6 @@
 use super::stacktrace::StackTrace;
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
-use std::collections::HashMap;
 
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize, JsonSchema)]
 pub struct ErrorData {
@@ -65,30 +64,6 @@ pub struct ThreadData {
     pub stack: StackTrace,
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub state: Option<String>,
-}
-
-impl From<(String, Vec<crate::StackFrame>)> for ThreadData {
-    fn from(value: (String, Vec<crate::StackFrame>)) -> Self {
-        let crashed = false; // Currently, only .Net uses this, and I believe they don't put the crashing thread here
-        let name = value.0;
-        let stack = value.1.into();
-        let state = None;
-        Self {
-            crashed,
-            name,
-            stack,
-            state,
-        }
-    }
-}
-
-pub fn thread_data_from_additional_stacktraces(
-    additional_stacktraces: HashMap<String, Vec<crate::StackFrame>>,
-) -> Vec<ThreadData> {
-    additional_stacktraces
-        .into_iter()
-        .map(|x| x.into())
-        .collect()
 }
 
 #[cfg(test)]
