+++ crashtracker/src/crash_info/proc_info.rs
@@ -8,12 +8,6 @@ pub struct ProcInfo {
     pub pid: u32,
 }
 
-impl From<crate::crash_info::ProcessInfo> for ProcInfo {
-    fn from(value: crate::crash_info::ProcessInfo) -> Self {
-        Self { pid: value.pid }
-    }
-}
-
 #[cfg(test)]
 impl super::test_utils::TestInstance for ProcInfo {
     fn test_instance(seed: u64) -> Self {
