+++ crashtracker/src/crash_info/spans.rs
