+++ crashtracker/src/crash_info/telemetry.rs
@@ -1,12 +1,10 @@
-// Copyright 2023-Present Datadog, Inc. https://www.datadoghq.com/
+// Copyright 2024-Present Datadog, Inc. https://www.datadoghq.com/
 // SPDX-License-Identifier: Apache-2.0
+use std::{fmt::Write, time::SystemTime};
 
-use std::collections::HashMap;
-use std::fmt::Write;
-use std::time::SystemTime;
-
-use super::{CrashInfo, CrashtrackerMetadata, StackFrame};
+use super::{CrashInfo, Metadata};
 use anyhow::{Context, Ok};
+use chrono::{DateTime, Utc};
 use ddcommon::Endpoint;
 use ddtelemetry::{
     build_host,
@@ -27,7 +25,7 @@ macro_rules! parse_tags {
             let mut $var: Option<&str> = None;
         )*
         for tag in $tag_iterator {
-            let Some((name, value)) = tag.as_ref().split_once(':') else {
+            let Some((name, value)) = tag.split_once(':') else {
                 continue;
             };
             match name {
@@ -39,19 +37,6 @@ macro_rules! parse_tags {
     };
 }
 
-#[derive(Debug, serde::Serialize)]
-/// This struct represents the part of the crash_info that we are sending in the
-/// log `message` field as a json
-struct TelemetryCrashInfoMessage<'a> {
-    pub additional_stacktraces: &'a HashMap<String, Vec<StackFrame>>,
-    pub files: &'a HashMap<String, Vec<String>>,
-    pub metadata: Option<&'a CrashtrackerMetadata>,
-    pub os_info: &'a os_info::Info,
-    pub span_ids: &'a Vec<u128>,
-    pub tags: &'a HashMap<String, String>,
-    pub trace_ids: &'a Vec<u128>,
-}
-
 pub struct TelemetryCrashUploader {
     metadata: TelemetryMetadata,
     cfg: ddtelemetry::config::Config,
@@ -59,7 +44,7 @@ pub struct TelemetryCrashUploader {
 
 impl TelemetryCrashUploader {
     pub fn new(
-        prof_metadata: &CrashtrackerMetadata,
+        crashtracker_metadata: &Metadata,
         endpoint: &Option<Endpoint>,
     ) -> anyhow::Result<Self> {
         let mut cfg = ddtelemetry::config::Config::from_env();
@@ -81,15 +66,15 @@ impl TelemetryCrashUploader {
         }
 
         parse_tags!(
-            prof_metadata.tags.iter(),
-            "service" => service_name,
-            "service_version" => service_version,
+            crashtracker_metadata.tags.iter(),
+            "env" => env,
             "language" => language_name,
-            "runtime_version" => language_version,
             "library_version" => library_version,
             "profiler_version" => profiler_version,
+            "runtime_version" => language_version,
             "runtime-id" => runtime_id,
-            "env" => env,
+            "service_version" => service_version,
+            "service" => service_name,
         );
 
         let application = Application {
@@ -121,26 +106,20 @@ impl TelemetryCrashUploader {
     pub async fn upload_to_telemetry(&self, crash_info: &CrashInfo) -> anyhow::Result<()> {
         let metadata = &self.metadata;
 
-        let message = serde_json::to_string(&TelemetryCrashInfoMessage {
-            additional_stacktraces: &crash_info.additional_stacktraces,
-            files: &crash_info.files,
-            metadata: crash_info.metadata.as_ref(),
-            os_info: &crash_info.os_info,
-            span_ids: &crash_info.span_ids,
-            tags: &crash_info.tags,
-            trace_ids: &crash_info.trace_ids,
-        })?;
+        let message = serde_json::to_string(crash_info)?;
 
-        let stack_trace = serde_json::to_string(&crash_info.stacktrace)?;
+        let stack_trace = serde_json::to_string(&crash_info.error.stack)?;
         let tags = extract_crash_info_tags(crash_info).unwrap_or_default();
 
-        let tracer_time = match &crash_info.timestamp {
-            Some(ts) => ts.timestamp() as u64,
-            None => SystemTime::UNIX_EPOCH
-                .elapsed()
-                .map(|d| d.as_secs())
-                .unwrap_or(0),
-        };
+        let tracer_time = crash_info.timestamp.parse::<DateTime<Utc>>().map_or_else(
+            |_| {
+                SystemTime::now()
+                    .duration_since(SystemTime::UNIX_EPOCH)
+                    .map(|d| d.as_secs())
+                    .unwrap_or(0)
+            },
+            |ts| ts.timestamp() as u64,
+        );
 
         let payload = data::Telemetry {
             tracer_time,
@@ -149,7 +128,6 @@ impl TelemetryCrashUploader {
             seq_id: 1,
             application: &metadata.application,
             host: &metadata.host,
-            origin: Some("Crashtracker"),
             payload: &data::Payload::Logs(vec![data::Log {
                 message,
                 level: LogLevel::Error,
@@ -158,6 +136,7 @@ impl TelemetryCrashUploader {
                 is_sensitive: true,
                 count: 1,
             }]),
+            origin: Some("Crashtracker"),
         };
         let client = ddtelemetry::worker::http_client::from_config(&self.cfg);
         let req = request_builder(&self.cfg)?
@@ -194,67 +173,66 @@ impl TelemetryCrashUploader {
 
 fn extract_crash_info_tags(crash_info: &CrashInfo) -> anyhow::Result<String> {
     let mut tags = String::new();
-    write!(&mut tags, "uuid:{}", crash_info.uuid)?;
-    if let Some(siginfo) = &crash_info.siginfo {
-        write!(&mut tags, ",signum:{}", siginfo.signum)?;
-        if let Some(signame) = &siginfo.signame {
-            write!(&mut tags, ",signame:{}", signame)?;
-        }
-        if let Some(faulting_address) = &siginfo.faulting_address {
-            write!(&mut tags, ",faulting_address:{:#018x}", faulting_address)?;
-        }
+    write!(
+        &mut tags,
+        "data_schema_version:{}",
+        crash_info.data_schema_version
+    )?;
+    if let Some(fingerprint) = &crash_info.fingerprint {
+        write!(&mut tags, ",fingerprint:{fingerprint}")?;
     }
+    write!(&mut tags, ",incomplete:{}", crash_info.incomplete)?;
+    write!(&mut tags, ",is_crash:{}", crash_info.error.is_crash)?;
+    write!(&mut tags, ",uuid:{}", crash_info.uuid)?;
     for (counter, value) in &crash_info.counters {
-        write!(&mut tags, ",{}:{}", counter, value)?;
+        write!(&mut tags, ",{counter}:{value}")?;
+    }
+
+    if let Some(siginfo) = &crash_info.sig_info {
+        if let Some(si_addr) = &siginfo.si_addr {
+            write!(&mut tags, ",si_addr:{si_addr}")?;
+        }
+        write!(&mut tags, ",si_code:{}", siginfo.si_code)?;
+        write!(
+            &mut tags,
+            ",si_code_human_readable:{:?}",
+            siginfo.si_code_human_readable
+        )?;
+        write!(&mut tags, ",si_signo:{}", siginfo.si_signo)?;
+        write!(
+            &mut tags,
+            ",si_signo_human_readable:{:?}",
+            siginfo.si_signo_human_readable
+        )?;
     }
     Ok(tags)
 }
+
 #[cfg(test)]
 mod tests {
-    use std::{
-        collections::{HashMap, HashSet},
-        fs,
-    };
-
-    use crate::SigInfo;
-    use chrono::DateTime;
-    use ddcommon::{tag, Endpoint};
-
     use super::TelemetryCrashUploader;
+    use crate::crash_info::{test_utils::TestInstance, CrashInfo, Metadata, StackTrace};
+    use ddcommon::Endpoint;
+    use std::{collections::HashSet, fs};
 
-    fn new_test_uploader() -> TelemetryCrashUploader {
+    fn new_test_uploader(seed: u64) -> TelemetryCrashUploader {
         TelemetryCrashUploader::new(
-            &new_test_prof_metadata(),
+            &Metadata::test_instance(seed),
             &Some(Endpoint::from_slice("http://localhost:8126")),
         )
         .unwrap()
     }
 
-    fn new_test_prof_metadata() -> super::CrashtrackerMetadata {
-        super::CrashtrackerMetadata {
-            library_name: "libdatadog".to_owned(),
-            library_version: "1.0.0".to_owned(),
-            family: "native".to_owned(),
-            tags: vec![
-                tag!("service", "foo"),
-                tag!("service_version", "bar"),
-                tag!("runtime-id", "xyz"),
-                tag!("language", "native"),
-            ],
-        }
-    }
-
     #[test]
     #[cfg_attr(miri, ignore)]
     fn test_profiler_config_extraction() {
-        let t = new_test_uploader();
+        let t = new_test_uploader(1);
 
         let metadata = t.metadata;
         assert_eq!(metadata.application.service_name, "foo");
         assert_eq!(metadata.application.service_version.as_deref(), Some("bar"));
         assert_eq!(metadata.application.language_name, "native");
         assert_eq!(metadata.runtime_id, "xyz");
-
         let cfg = t.cfg;
         assert_eq!(
             cfg.endpoint.unwrap().url.to_string(),
@@ -264,50 +242,32 @@ mod tests {
 
     #[tokio::test]
     #[cfg_attr(miri, ignore)]
-    async fn test_crash_request_content() {
+    async fn test_crash_request_content() -> anyhow::Result<()> {
+        // This keeps alive for scope
         let tmp = tempfile::tempdir().unwrap();
         let output_filename = {
             let mut p = tmp.into_path();
             p.push("crash_info");
             p
         };
-        let mut t = new_test_uploader();
+        let seed = 1;
+        let mut t = new_test_uploader(seed);
 
         t.cfg
             .set_host_from_url(&format!("file://{}", output_filename.to_str().unwrap()))
             .unwrap();
+        let test_instance = super::CrashInfo::test_instance(seed);
 
-        let mut counters = HashMap::new();
-        counters.insert("collecting_sample".to_owned(), 1);
-        counters.insert("not_profiling".to_owned(), 0);
-        t.upload_to_telemetry(&crate::CrashInfo {
-            counters,
-            files: HashMap::new(),
-            metadata: Some(new_test_prof_metadata()),
-            os_info: os_info::Info::unknown(),
-            siginfo: Some(SigInfo {
-                signum: 11,
-                signame: Some("SIGSEGV".to_owned()),
-                faulting_address: Some(0x1234),
-            }),
-            proc_info: None,
-            stacktrace: vec![],
-            span_ids: vec![42, 24],
-            trace_ids: vec![345, 666],
-            additional_stacktraces: HashMap::new(),
-            tags: HashMap::new(),
-            timestamp: DateTime::from_timestamp(1702465105, 0),
-            uuid: uuid::uuid!("1d6b97cb-968c-40c9-af6e-e4b4d71e8781"),
-            incomplete: true,
-        })
-        .await
-        .unwrap();
+        t.upload_to_telemetry(&test_instance).await.unwrap();
 
         let payload: serde_json::value::Value =
             serde_json::de::from_str(&fs::read_to_string(&output_filename).unwrap()).unwrap();
-        assert_eq!(payload["request_type"], "logs");
-        assert_eq!(payload["application"]["service_name"], "foo");
+        assert_eq!(payload["api_version"], "v2");
         assert_eq!(payload["application"]["language_name"], "native");
+        assert_eq!(payload["application"]["service_name"], "foo");
+        assert_eq!(payload["application"]["service_version"], "bar");
+        assert_eq!(payload["request_type"], "logs");
+        assert_eq!(payload["tracer_time"], 1568898000);
 
         assert_eq!(payload["payload"].as_array().unwrap().len(), 1);
         let tags = payload["payload"][0]["tags"]
@@ -317,15 +277,28 @@ mod tests {
             .collect::<HashSet<_>>();
         assert_eq!(
             HashSet::from_iter([
-                "uuid:1d6b97cb-968c-40c9-af6e-e4b4d71e8781",
-                "signum:11",
-                "signame:SIGSEGV",
                 "collecting_sample:1",
+                "data_schema_version:1.0",
+                "incomplete:true",
+                "is_crash:true",
                 "not_profiling:0",
-                "faulting_address:0x0000000000001234",
+                "si_addr:0x0000000000001234",
+                "si_code_human_readable:SEGV_BNDERR",
+                "si_code:1",
+                "si_signo_human_readable:SIGSEGV",
+                "si_signo:11",
+                "uuid:1d6b97cb-968c-40c9-af6e-e4b4d71e8781",
             ]),
             tags
         );
         assert_eq!(payload["payload"][0]["is_sensitive"], true);
+        assert_eq!(payload["payload"][0]["level"], "ERROR");
+        let stack_trace: StackTrace =
+            serde_json::from_str(payload["payload"][0]["stack_trace"].as_str().unwrap())?;
+        assert_eq!(stack_trace, test_instance.error.stack);
+        let body: CrashInfo =
+            serde_json::from_str(payload["payload"][0]["message"].as_str().unwrap())?;
+        assert_eq!(body, test_instance);
+        Ok(())
     }
 }
